@rtfb 2017-07-27
The fix looks good. Would you also be willing to forward port it to v2? The internals of Smartypants haven't changed much there, IIRC.