@bep 2017-07-27
>The fix looks good. Would you also be willing to forward port it to v2?

No problem, I will do that in a separate PR. I created #380 to track/remind me of it.